[Chore] 포톤매니저 교환
