+++ Assets/JWH/JWH_Scrips/PhotonManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cc433f626bac06f46a4754a0ab09b462
+guid: 23e2b8689c2810b4790c4443a2afdb2f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
