--- Assets/JWH/JWH_Scrips/RealPhotonManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 23e2b8689c2810b4790c4443a2afdb2f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
