--- Assets/JWH/JWH_Scrips/TestRoomSceneManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d271c5d6d622096439c3938357e6248c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
